Bash 3.1 patch 010
Application of bash-3.1-patches/bash31-010.

This is the second release of the bash 3.1 patch 10, which was updated
due to a serious bug. It is included here despite being chronologically
after patch 11, as the diff applies cleanly here.

It is unclear what system maintainers who had applied patch 11 were
meant to do here.

Please file an issue if you can provide more insight into the better way
in which to do order this.