Changes made prior to 1.14.7 (unlinked)
These changes are all included in a diff between 1.14.6 in this
repository and the 1.14.7 tarball, however, not in the upstream
bash-1.14.6-1.14.7.diff.gz.

This means these changes occured at some point in the patch series used
in this repo that was used to get to this point, but it is not clear
which patch has these changes missing.

Please file an issue in this repository if you can find which patch
these should be linked to.